Arreglado errores menores de compilacion

+++ funciones_tp2.c
@@ -30,7 +30,7 @@ void atender_siguiente_paciente(doctor_t* doctor, hash_t* lista_pacientes_totale
 
 void informe_doctores(char letra_inicio, char letra_final, abb_t* doctores){
     abb_iter_t* iterador = abb_iter_in_crear(doctores);
-    int contador = 0;
+    size_t contador = 0;
     while(!abb_iter_in_al_final(iterador)){
         doctor_t* doctor_actual = (doctor_t*)abb_iter_in_ver_actual(iterador);
         if(strcmp(doctor_actual->nombre, letra_inicio) < 0){
